Fixed suboptimal tokenS spending estimation in P2P rings

+++ contracts/impl/Data.sol
@@ -70,7 +70,7 @@ library Data {
         IFeeHolder feeHolder;
         uint64 ringIndex;
         Tax tax;
-        uint16 feePercentageBase;
+        uint feePercentageBase;
     }
 
     struct Mining {
